@@ -4,34 +4,27 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Logging;
 using GitServer.Infrastructure;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Linq;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Authentication.Cookies;
-using System.Security.Claims;
 using GitServer.Handlers;
 using GitServer.ApplicationCore.Interfaces;
 using NetCoreBBS.Infrastructure.Repositorys;
 using GitServer.ApplicationCore.Models;
+using Microsoft.Extensions.Hosting;
 
 namespace GitServer
 {
     public class Startup
     {
-        public Startup(IHostingEnvironment env)
+        public Startup(IConfiguration configuration)
         {
-            var builder = new ConfigurationBuilder()
-                .SetBasePath(env.ContentRootPath)
-                .AddJsonFile("appsettings.json", optional: false, reloadOnChange: true)
-                .AddJsonFile($"appsettings.{env.EnvironmentName}.json", optional: true)
-                .AddEnvironmentVariables();
-            Configuration = builder.Build();
+            Configuration = configuration;
         }
 
-        public IConfigurationRoot Configuration { get; }
+        public IConfiguration Configuration { get; }
 
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
@@ -54,7 +47,9 @@ namespace GitServer
                     break;
             }
             // Add framework services.
-            services.AddMvc();
+            services.AddMvc(option=> {
+                option.EnableEndpointRouting = false;
+            });
 			services.AddOptions();
 
             services.AddAuthentication(options =>
@@ -76,10 +71,8 @@ namespace GitServer
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env, ILoggerFactory loggerFactory)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-            loggerFactory.AddConsole(Configuration.GetSection("Logging"));
-            loggerFactory.AddDebug();
             InitializeGitServerDatabase(app.ApplicationServices);
 			if(env.IsDevelopment())
 			{
